# `math`: Optimize `Distance`, `Clamp`, `Approach`