`Distance`: `math.sqrt( ... )` —> `( ... ) ^ 0.5`

![image](https://github.com/user-attachments/assets/5a357559-88f6-48ac-880d-c9ceca4c215c)
<details> <summary>Benchmark Code</summary>

```lua

times = 64

print( Format( 'branch => %s\niterations => %i', BRANCH, times ) )

function main()

	print( Format( '\njit status => %s', jit.status() ) )

	sum = 0
	avg1 = 0

	local sqrt = math.sqrt

	for i = 1, times do

		util.TimerCycle()

			local x = sqrt( 5 )

		sum = sum + util.TimerCycle()

	end

	avg1 = sum / times
	print( Format( 'math.sqrt\n\t%.8f ms on average (100%%)', sum / times ) )

	sum = 0
	avg2 = 0

	for i = 1, times do

		util.TimerCycle()

			local x = 5 ^ 0.5

		sum = sum + util.TimerCycle()

	end

	avg2 = sum / times
	print( Format( 'pow operator\n\t%.8f ms on average (%.1f%%)', sum / times, ( avg2 * 100 ) / avg1 ) )

end

jit.off()
main()

jit.on()
main()

```
</details>

`Clamp` & `Approach`: Replaced functions with operators, which are faster in action.

![image](https://github.com/user-attachments/assets/e072b5cc-fb6d-4d10-b565-a60e3ebf6229)
![image](https://github.com/user-attachments/assets/3d277645-50e7-4b98-9c36-b434d8314fd3)

Note: Approach becomes a bit slower with enabled JIT if used more than once at the moment (applies to the default branches, but not to `x86-64`)