@Astralcircle 2025-02-23
New math.Clamp function does not handle nan properly:
```lua
print(math.Clamp(0 / 0, 1, 5), math.Clamp2(0 / 0, 1, 5))
```
```
1	nan
```