@goodusername123 2025-02-27
replacing `math.sqrt` with `^ 0.5` seems like a really really bad idea future proofing wise considering that in modern versions of LuaJIT the performance characteristics of these are somewhat wildly different then they are in the versions of LuaJIT GMod currently uses.
https://www.freelists.org/post/luajit/Bad-performance-of-x05-versus-mathsqrtx
https://github.com/LuaJIT/LuaJIT/issues/684#issuecomment-1020116837
https://github.com/Facepunch/garrysmod/pull/2198#issuecomment-2626726754