@noaccessl 2025-03-02
Ok, I have figured out some balanced optimization for `math.Clamp`. Decided to replace two conditions with one with `elseif`-element. Also, moved NaN-check to the first condition. 

# Previous
```lua
function math.Clamp( num, low, high )
	if ( num < low ) then return low end
	if ( num > high ) then return high end

	-- If the passed number is NaN, return the lowest limit
	return num ~= num and low or num
end
```
luac.nl
![image](https://github.com/user-attachments/assets/b489b33d-183c-4ba3-bbcf-3ad8a2263e2a)
luajit.me
![image](https://github.com/user-attachments/assets/87b82000-f5cb-46f4-83e2-d992f76dbeec)

# New
```lua
function math.Clamp( num, low, high )
	if ( num < low or num ~= num --[[NaN-check]] ) then
		return low
	elseif ( num > high ) then
		return high
	end

	return num
end
```
luac.nl
![image](https://github.com/user-attachments/assets/0cff82b4-97b3-4487-ac35-5ea6e0bbe865)
luajit.me
![image](https://github.com/user-attachments/assets/7e34e877-0d74-4413-a48c-90516732a6cd)

# Benchmark
[Benchmark I use](https://gist.github.com/noaccessl/2336bb390aca15bc7ba50d2c9d966daa)

<details> <summary>Benchmark code</summary>

```lua
local min, max = math.min, math.max

local function Clamp_mathlib( num, low, high )

	return min( max( num, low ), high )

end

local function Clamp_prev( num, low, high )

	if ( num < low ) then return low end
	if ( num > high ) then return high end

	-- If the passed number is NaN, return the lowest limit
	return num ~= num and low or num

end

local function Clamp_new( num, low, high )

	if ( num < low or num ~= num --[[NaN-check]] ) then
		return low
	elseif ( num > high ) then
		return high
	end

	return num

end

local Tests = {

	'Clamp_mathlib', function()

		Clamp_mathlib( -1, 0, 5 )
		Clamp_mathlib( 0 / 0, 0, 5 )
		Clamp_mathlib( 10, 0, 5 )
		Clamp_mathlib( 3, 0, 5 )

	end;

	'Clamp_prev', function()

		Clamp_prev( -1, 0, 5 )
		Clamp_prev( 0 / 0, 0, 5 )
		Clamp_prev( 10, 0, 5 )
		Clamp_prev( 3, 0, 5 )

	end;

	'Clamp_new', function()

		Clamp_new( -1, 0, 5 )
		Clamp_new( 0 / 0, 0, 5 )
		Clamp_new( 10, 0, 5 )
		Clamp_new( 3, 0, 5 )

	end;

}

local Frames		= 1000
local Digit 		= 4
local Iterations	= 30

SetupBenchmark( Frames, Digit, Iterations, 'us', 'average' )

local function RunTests()

	for i = 1, #Tests, 2 do
		Benchmark( Tests[i], Tests[i + 1], i == 1 --[[The first test will be considered as standard]] )
	end

end

jit.off()
RunTests()

jit.on()
RunTests()
```
</details>

# Results
## Pre-Release Branch
![image](https://github.com/user-attachments/assets/4d6055e7-7828-4084-85d5-684ad1e13a44)

## x86-64 Branch
![image](https://github.com/user-attachments/assets/bfcde03e-c7fb-4798-b92d-4403e7bd9e07)

---

Waiting for feedback.