@IVogel 2025-03-03
Used same benchmark as you, and your parameters and got these results :/

_Windows 10 x86_64, Ryzen 5 3600_
 Dev branch  |  x86-64 branch, 64bit  |  x86-64 branch, 32bit 
:--------------:|:-------------------------:|:-------------------------:
![benchmark-dev](https://github.com/user-attachments/assets/2ec0a8d6-70ab-452c-a9d0-307bf9212f50) | ![benchmark-x86-64-64](https://github.com/user-attachments/assets/432268d5-58fd-4850-81cc-be7bccb57599) | ![benchmark-x86-64-32](https://github.com/user-attachments/assets/14f3f174-6788-4c82-ac52-08421808ead2)