@JarosLucky 2025-03-03
Windows 10 x86_64, Intel I5 3550
x86-64 branch, 64bit
![image](https://github.com/user-attachments/assets/a595209a-e6dd-45af-b2d5-9a1be8428378)
Since the x86-64 branch uses the updated LuaJIT with Stitching, I don't care about the results with JIT disabled as a measure of efficiency. After all, nobody uses JIT disabling besides special cases, besides functions that could not compile with JIT