@JarosLucky 2025-03-03
If you look at your own benchmark result, you can see that the original feature is already working quite well. It makes no sense to optimize the version without jit when regressing performance with jit. **My opinion this change should definitely be brought back. And the others are measured again, as the benchmark attached originally has questionable validity**