@JarosLucky 2025-03-03
> # Benchmark
> [Benchmark I use](https://gist.github.com/noaccessl/2336bb390aca15bc7ba50d2c9d966daa)

Also, it seems you need to rewrite your benchmark to measure performance without JIT because the change in execution speed between with and without JIT differs by 76.67 times