@noaccessl 2025-03-10
Ok, I've reworked my benchmark. 

`math.sqrt( x )` is more stable in speed than `x ^ 0.5` (especially on the `x86-64`). And the previous `math.Clamp` & `math.Approach` was handled by JIT well enough.

Overall, the initial functions are well-optimized for JIT; and the majority doesn't turn it off globally. Was a bit presumptuous with my optimizations as I was prioritizing the results with disabled JIT.